audiobot/pyproject.toml

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "audiobot"
version = "0.1.0"
dependencies = [
    "google-cloud-speech",
    "scipy",
    "google-generativeai",
    "fastapi",
    "uvicorn",
    "requests",
    "numpy",
    "pydantic",
    "tqdm",
    "google-api-python-client",
    "google-auth-httplib2",
    "httplib2",
    "uritemplate",
    "Jinja2",
    "python-multipart",
    # ML training / audio
    "torch>=2.1.0; platform_system != 'Darwin' or platform_machine != 'arm64'",
    "pytorch-lightning>=2.2.0",
    "torchaudio>=2.1.0; platform_system != 'Darwin' or platform_machine != 'arm64'",
    "soundfile",
    "librosa",
    # Vertex AI + GCS
    "google-cloud-aiplatform",
    "google-cloud-storage",
]

[tool.setuptools.packages.find]
where = ["."]
